Move RealTorController to internal package
