+++ library/controller/kmp-tor-controller/src/androidMain/kotlin/io/matthewnelson/kmp/tor/controller/TorController.kt
@@ -27,12 +27,12 @@ import io.matthewnelson.kmp.tor.controller.common.events.TorEventProcessor
 import io.matthewnelson.kmp.tor.controller.common.exceptions.TorControllerException
 import io.matthewnelson.kmp.tor.controller.common.file.Path
 import io.matthewnelson.kmp.tor.controller.common.internal.isUnixPath
+import io.matthewnelson.kmp.tor.controller.internal.controller.RealTorController
 import io.matthewnelson.kmp.tor.controller.internal.controller.getTorControllerDispatcher
 import io.matthewnelson.kmp.tor.controller.internal.io.ReaderWrapper
 import io.matthewnelson.kmp.tor.controller.internal.io.SocketWrapper
 import io.matthewnelson.kmp.tor.controller.internal.io.WriterWrapper
 import io.matthewnelson.kmp.tor.controller.internal.util.toTorController
-import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.withContext
 import java.io.*
 import java.net.Socket
@@ -43,7 +43,7 @@ import java.net.SocketException
  * asynchronous communication.
  *
  * Upon connecting, [TorController] will run continuously until
- * Tor has been shutdown.
+ * Tor has been shutdown or [disconnect] has been called.
  *
  * @see [newInstance]
  * @see [RealTorController]
@@ -107,8 +107,7 @@ actual interface TorController: TorControlProcessor, TorEventProcessor<TorEvent.
                     )
                 }
 
-                @OptIn(ExperimentalCoroutinesApi::class)
-                return realTorController(
+                return RealTorController(
                     reader = ReaderWrapper.wrap(reader),
                     writer = WriterWrapper.wrap(writer),
                     socket = SocketWrapper.wrap(socket),
